Move backend tests to routes/specs
